+++ prettier.config.js
@@ -1,0 +1,2 @@
+
+export const plugins = [require('prettier-plugin-tailwindcss')];
