+++ src/App.tsx
@@ -1,14 +1,10 @@
-
 import "./App.css"
 import { BrowserRouter, Routes, Route } from "react-router-dom"
 
 //components
 import Navbar from "./comps/navbar"
-
-
 import LandingPage from "./page/landingPage"
 import Homepage from "./page/Homepage/Homepage"
-
 
 function App() {
    return (
