+++ src/page/MainGrid/MainGrid.tsx
@@ -1,35 +1,9 @@
 import MainGridItem from "./MainGridItem";
-import { useEffect, useState } from "react";
-
+import { useFetch } from "../../hooks/useFetch";
 
 const MainGrid = () => {
-
-const [data, setData] = useState<any>()
-
-// fetch mock pics here with a hook
-
-useEffect(() => {
-    const fetchData = async () => {
-        fetch('http://localhost:3000/posts')
-        .then(res => res.json())
-        .then(data => {
-            console.log(data);
-
-            if(data.error) {
-                console.log(data.error);
-            }
-            else {
-                console.log('succes!');
-                setData(data)
-            }
-        })
-    }
+    const {fetchData, data} = useFetch('http://localhost:3000/posts')
     fetchData()
-},[])
-
-    console.log(data);
-    
-
 
     return (  
         <div className="grid relative gap-x-3 m-2 gap-y-10 grid-cols-2">
