+++ src/page/MainGrid/MainGridItem.tsx
@@ -9,7 +9,8 @@ const MainGridItem = ({image} : {image:string}) => {
             <p>item descriptions</p>
             <p>prix : 100 eurodollars</p>
           </div>
-        <div className='absolute top-0 left-0 right-0 bottom-0 h-[14rem] bg-black opacity-0 hover:opacity-10 hover:transition-all duration-300 ease-in'></div>
+        <div className='absolute top-0 left-0 right-0 bottom-0 h-[14rem] bg-black opacity-0 hover:opacity-10 
+        hover:transition-all duration-300 ease-in'></div>
         </div>
     );
   }
