+++ src/vite-env.d.ts
@@ -1,1 +1,3 @@
 /// <reference types="vite/client" />
+
+declare module 'prettier-plugin-tailwindcss';
